+++ packages/maps/jest.config.js
@@ -1,4 +1,8 @@
 const { BaseJestConfig } = require('../../config/jest');
 module.exports = {
   ...BaseJestConfig,
+  moduleNameMapper: {
+    '^lodash-es$': 'lodash',
+    '^.+\\.(css|less)$': 'identity-obj-proxy',
+  },
 };